refactor: json response

+++ controllers/scholarship.controller.ts
@@ -1,4 +1,4 @@
-import { PrismaClient, Role } from "@prisma/client";
+import { PrismaClient } from "@prisma/client";
 import { Request, Response } from "express";
 
 const prisma = new PrismaClient();
@@ -16,6 +16,23 @@ export const createScholarship = async (req: Request, res: Response) => {
             type,
         } = req.body;
 
+        const existingOrganization = await prisma.organization.findUnique({
+            select: {
+                user :{
+                    select: {
+                        name: true,
+                    },
+                },
+            },
+            where: {
+                organization_id,
+            },
+        });
+
+        if (!existingOrganization) {
+            throw new Error("Organization not found");
+        }
+
         const scholarship = await prisma.scholarship.create({
             data: {
                 title,
@@ -24,11 +41,7 @@ export const createScholarship = async (req: Request, res: Response) => {
                 coverage,
                 contact_name,
                 contact_email,
-                organization: {
-                    connect: {
-                        organization_id,
-                    },
-                },
+                organization_id,
                 scholarshiptype: {
                     create: type.map((element: string) => ({
                         type: element,
@@ -40,9 +53,24 @@ export const createScholarship = async (req: Request, res: Response) => {
             },
         });
 
-        res.status(201).json(scholarship);
+        res.status(201).json({
+            status: "success",
+            message: "Scholarship created successfully",
+            data: {
+                organization_id: scholarship.organization_id,
+                scholarship_id: scholarship.scholarship_id,
+                title: scholarship.title,
+                description: scholarship.description,
+                short_description: scholarship.short_description,
+                coverage: scholarship.coverage,
+                contact_name: scholarship.contact_name,
+                contact_email: scholarship.contact_email,
+                type: scholarship.scholarshiptype,
+            }
+        });
     } catch (error: any) {
         res.status(500).json({
+            status: "error",
             message: error.message,
         });
     }
@@ -53,13 +81,33 @@ export const getScholarships = async (req: Request, res: Response) => {
         const scholarships = await prisma.scholarship.findMany({
             include: {
                 scholarshiptype: true,
-                organization: true,
             },
         });
 
-        res.status(200).json(scholarships);
+        if (!scholarships) {
+            throw new Error("Scholarships not found");
+        }
+
+        res.status(200).json({
+            status: "success",
+            message: "Scholarships retrieved successfully",
+            data: scholarships.map((scholarship) => {
+                return {
+                    organization_id: scholarship.organization_id,
+                    scholarship_id: scholarship.scholarship_id,
+                    title: scholarship.title,
+                    description: scholarship.description,
+                    short_description: scholarship.short_description,
+                    coverage: scholarship.coverage,
+                    contact_name: scholarship.contact_name,
+                    contact_email: scholarship.contact_email,
+                    type: scholarship.scholarshiptype,
+                };
+            }),
+        });
     } catch (error: any) {
         res.status(500).json({
+            status: "error",
             message: error.message,
         });
     }
@@ -76,9 +124,29 @@ export const getScholarship = async (req: Request, res: Response) => {
                 organization: true,
             },
         });
-        res.status(200).json(scholarship);
+
+        if (!scholarship) {
+            throw new Error("Scholarship not found");
+        }
+
+        res.status(200).json({
+            status: "success",
+            message: "Scholarship retrieved successfully",
+            data: {
+                organization_id: scholarship.organization_id,
+                scholarship_id: scholarship.scholarship_id,
+                title: scholarship.title,
+                description: scholarship.description,
+                short_description: scholarship.short_description,
+                coverage: scholarship.coverage,
+                contact_name: scholarship.contact_name,
+                contact_email: scholarship.contact_email,
+                type: scholarship.scholarshiptype,
+            },
+        });
     } catch (error: any) {
         res.status(500).json({
+            status: "error",
             message: error.message,
         });
     }
@@ -93,7 +161,6 @@ export const updateScholarship = async (req: Request, res: Response) => {
             coverage,
             contact_name,
             contact_email,
-            organization_id,
             type,
         } = req.body;
         const { id } = req.params;
@@ -122,11 +189,6 @@ export const updateScholarship = async (req: Request, res: Response) => {
                 coverage,
                 contact_name,
                 contact_email,
-                organization: {
-                    connect: {
-                        organization_id,
-                    },
-                },
                 scholarshiptype: {
                     deleteMany: {
                         scholarship_id: Number(id),
@@ -142,10 +204,24 @@ export const updateScholarship = async (req: Request, res: Response) => {
             },
         });
 
-        res.status(200).json(updatedScholarship);
+        res.status(200).json({
+            status: "success",
+            message: "Scholarship updated successfully",
+            data: {
+                organization_id: updatedScholarship.organization_id,
+                scholarship_id: updatedScholarship.scholarship_id,
+                title: updatedScholarship.title,
+                description: updatedScholarship.description,
+                short_description: updatedScholarship.short_description,
+                coverage: updatedScholarship.coverage,
+                contact_name: updatedScholarship.contact_name,
+                contact_email: updatedScholarship.contact_email,
+                type: updatedScholarship.scholarshiptype,
+            },
+        });
     } catch (error: any) {
         res.status(500).json({
-            code: 500,
+            status: "error",
             message: error.message,
         });
     }
@@ -154,6 +230,21 @@ export const updateScholarship = async (req: Request, res: Response) => {
 export const deleteScholarship = async (req: Request, res: Response) => {
     try {
         const { id } = req.params;
+        
+        const existingScholarship = await prisma.scholarship.findUnique({
+            where: {
+                scholarship_id: Number(id),
+            },
+            include: {
+                scholarshiptype: true,
+                organization: true,
+            },
+        });
+
+        if (!existingScholarship) {
+            throw new Error("Scholarship not found");
+        }
+
         const scholarship = await prisma.scholarship.delete({
             where: {
                 scholarship_id: Number(id),
@@ -163,10 +254,25 @@ export const deleteScholarship = async (req: Request, res: Response) => {
                 organization: true,
             },
         });
-        res.status(200).json(scholarship);
+
+        res.status(200).json({
+            status: "success",
+            message: "Scholarship deleted successfully",
+            data: {
+                organization_id: scholarship.organization_id,
+                scholarship_id: scholarship.scholarship_id,
+                title: scholarship.title,
+                description: scholarship.description,
+                short_description: scholarship.short_description,
+                coverage: scholarship.coverage,
+                contact_name: scholarship.contact_name,
+                contact_email: scholarship.contact_email,
+                type: scholarship.scholarshiptype,
+            },
+        });
     } catch (error: any) {
         res.status(500).json({
-            code: 500,
+            status: "error",
             message: error.message,
         });
     }
